+++ Rdmp.Dicom/Cache/Pipeline/PACSSource.cs
@@ -66,11 +66,16 @@ namespace Rdmp.Dicom.Cache.Pipeline
 
         [DemandsInitialization("Ignore whitelist of patient identifiers",defaultValue: false, mandatory: true)]
         public bool IgnoreWhiteList { get; set; }
-
-        [DemandsInitialization("Set the DICOM priority (recommended value for NHS transfers is low)", defaultValue: DicomPriority.Low, mandatory: true)]
-        public DicomPriority Priority { get; set; }
+        
+        /// <summary>
+        /// The maximum number of tries to fetch a given Study from the PACS.  Note that retries requests might not be issued immediately after
+        /// a failure.
+        /// </summary>
+        [DemandsInitialization("Maximum number of times to re-request a Study when a Failure is encountered",defaultValue:3 , mandatory: true)]
+        public int MaxRetries { get; set; } = 3;
 
         private HashSet<string> _whitelist;
+
 
         public override SMIDataChunk DoGetChunk(ICacheFetchRequest cacheRequest, IDataLoadEventListener listener,GracefulCancellationToken cancellationToken)
         {
@@ -81,7 +86,6 @@ namespace Rdmp.Dicom.Cache.Pipeline
             var requestSender = new DicomRequestSender(dicomConfiguration, listener);
             var dateFrom = Request.Start;
             var dateTo = Request.End;
-            var hasTransferTimedOut = false;
             CachingSCP.LocalAet = LocalAETitle;
             CachingSCP.Listener = listener;
 
@@ -99,22 +103,10 @@ namespace Rdmp.Dicom.Cache.Pipeline
                 Layout = cacheLayout
             };
             
-            ConcurrentBag<string> studiesToOrder = new ConcurrentBag<string>();
-
-            //enforce a minimum timeout
-            var transferTimeOutTimer = new Timer(Math.Max(30000,dicomConfiguration.TransferTimeOutInMilliseconds));
-            transferTimeOutTimer.Elapsed += (source, eventArgs) =>
-            {
-                hasTransferTimedOut = true;
-                listener.OnNotify(this,
-                new NotifyEventArgs(ProgressEventType.Information, "Transfer Timeout Exception Generated"));
-                throw new TimeoutException("Transfer Timeout Exception");
-            };
+            ConcurrentBag<StudyToFetch> studiesToOrder = new ConcurrentBag<StudyToFetch>();
 
             CachingSCP.OnEndProcessingCStoreRequest = (storeRequest, storeResponse) =>
-            {
-                transferTimeOutTimer.Reset();
-                
+            {                
                 SaveSopInstance(storeRequest,cacheLayout,listener);
                 listener.OnNotify(this,
                     new NotifyEventArgs(ProgressEventType.Debug,
@@ -139,7 +131,7 @@ namespace Rdmp.Dicom.Cache.Pipeline
                     request.OnResponseReceived += (req, response) =>
                     {
                         if (Filter(_whitelist, response))
-                            studiesToOrder.Add(response.Dataset.GetSingleValue<string>(DicomTag.StudyInstanceUID));
+                            studiesToOrder.Add(new StudyToFetch(response.Dataset.GetSingleValue<string>(DicomTag.StudyInstanceUID)));
 
                     };
                     requestSender.ThrottleRequest(request,client, cancellationToken.AbortToken);
@@ -153,10 +145,11 @@ namespace Rdmp.Dicom.Cache.Pipeline
 
                     var transferStopwatch = new Stopwatch();
 
-                    string current;
+                    StudyToFetch current;
+                    int consecutiveFailures = 0;
                                             
                     //While we have things to fetch
-                    while(studiesToOrder.TryTake(out current) && !hasTransferTimedOut)
+                    while(studiesToOrder.TryTake(out current))
                     {
                         transferStopwatch.Restart();
                         //delay value in mills
@@ -171,54 +164,84 @@ namespace Rdmp.Dicom.Cache.Pipeline
                         bool done = false;
 
                         //Build fetch command that Study
-                        var cMoveRequest = CreateCMoveByStudyUid(LocalAETitle,current, listener);
+                        var cMoveRequest = CreateCMoveByStudyUid(LocalAETitle,current.StudyUid, listener);
 
                         //Register callbacks
                         cMoveRequest.OnResponseReceived += (requ, response) =>
                         {
-                            if (response.Status.State == DicomState.Pending)
-                            {
-                                listener.OnNotify(this,
-                                    new NotifyEventArgs(ProgressEventType.Debug,
-                                        "Request: " + requ.ToString() + "items remaining: " + response.Remaining));
-                            }
-                            else if (response.Status.State == DicomState.Success)
-                            {
-                                listener.OnNotify(this,
-                                    new NotifyEventArgs(ProgressEventType.Debug,
-                                        "Request: " + requ.ToString() + "completed successfully"));
+                            listener.OnNotify(this,
+                                new NotifyEventArgs(ProgressEventType.Debug,
+                                $"Got {response.Status.State} response for {requ}.  Items remaining {response.Remaining}"));
 
-                                done = true;
-                            }
-                            else if (response.Status.State == DicomState.Failure)
+                            switch(response.Status.State)
                             {
-                                listener.OnNotify(this,
-                                    new NotifyEventArgs(ProgressEventType.Debug,
-                                        "Request: " + requ.ToString() + "failed to download: " + response.Failures));
+                                case DicomState.Pending : 
+                                case DicomState.Warning : 
+                                        // ignore
+                                        break;                                 
+                                case DicomState.Cancel : 
+                                case DicomState.Failure :
+                                    consecutiveFailures++;
                                     
-                                //if we can't get the study don't sit waiting for it to finish!
-                                done = true;
+                                    if(current.RetryCount < MaxRetries)
+                                    {
+                                        // put it back in the bag with a increased retry count
+                                        current.RetryCount++;
+                                        studiesToOrder.Add(current);
+                                    }
+
+                                    // final state
+                                    done = true;
+                                    break;
+                                case DicomState.Success :
+                                    // final state
+                                    consecutiveFailures = 0;
+                                    done = true;
+                                    break;
                             }
                         };
                         
                         //send the command to the server
+                        
+                        //tell user what we are sending
+                        listener.OnNotify(this,new NotifyEventArgs(ProgressEventType.Information,CMoveRequestToString(cMoveRequest,current.RetryCount +1)));
 
                         //do not use requestSender.ThrottleRequest(cMoveRequest, cancellationToken);
                         //TODO is there any need to throtttle this request given its lifetime
                         requestSender.ThrottleRequest(cMoveRequest, client, cancellationToken.AbortToken);
                                                 
-                        transferTimeOutTimer.Reset();
+                        
+                        //enforce a minimum timeout
+                        var swStudyTransfer = Stopwatch.StartNew();
+                        bool hasTransferTimedOut = false;
 
                         do
                         {
                             Task.Delay(Math.Max(100,dicomConfiguration.TransferPollingInMilliseconds), cancellationToken.AbortToken)
                                 .Wait(cancellationToken.AbortToken);
+                            
+                           hasTransferTimedOut = swStudyTransfer.ElapsedMilliseconds > dicomConfiguration.TransferTimeOutInMilliseconds;
                                 
                         }while(!done && !hasTransferTimedOut);
 
-                        transferTimeOutTimer.Stop();
-                        listener.OnNotify(this,
-                            new NotifyEventArgs(ProgressEventType.Information,CMoveRequestToString(cMoveRequest,1)));
+                        // Study has finished being fetched (or timed out)
+                        
+                        if(hasTransferTimedOut)
+                            listener.OnNotify(this,new NotifyEventArgs(ProgressEventType.Information,"Abandonning fetch of study " + current.StudyUid));
+                        
+                        if(consecutiveFailures > 5)
+                            throw new Exception("Too many consecutive failures, giving up");
+
+                        // 1 failure = study not available, 2 failures = system is having a bad day?
+                        if(consecutiveFailures > 1)
+                        {
+                            //wait 4 minutes then 6 minutes then 8 minutes, eventually server will start responding again?
+                            int sleepFor = consecutiveFailures * 2 * 60_000;
+                            listener.OnNotify(this,new NotifyEventArgs(ProgressEventType.Warning,$"Sleeping for {sleepFor}ms due to {consecutiveFailures} consecutive failures"));
+
+                            Task.Delay( sleepFor, cancellationToken.AbortToken)
+                                .Wait(cancellationToken.AbortToken);
+                        }
                     }
                         
                     #endregion
@@ -229,7 +252,6 @@ namespace Rdmp.Dicom.Cache.Pipeline
                 }
             }
 
-            transferTimeOutTimer.Dispose();
             return Chunk;
         }
 
@@ -377,8 +399,7 @@ namespace Rdmp.Dicom.Cache.Pipeline
                 RequestCooldownInMilliseconds = 1000 * RequestCooldownInSeconds,
                 TransferCooldownInMilliseconds = 1000 * TransferCooldownInSeconds,
                 TransferPollingInMilliseconds = 1000 * TransferPollingInSeconds,
-                TransferTimeOutInMilliseconds = 1000 * TransferTimeOutInSeconds,
-                Priority = Priority,
+                TransferTimeOutInMilliseconds = 1000 * TransferTimeOutInSeconds
             };
         }
         #endregion